update name everywhere
