+++ yarpc/abc.py
@@ -1,5 +1,5 @@
 """
-IOMirea-rpc - RPC system for IOMirea messenger
+yarpc - RPC system for IOMirea messenger
 Copyright (C) 2019  Eugene Ershov
 
 This program is free software: you can redistribute it and/or modify
